pyproject.toml
[tool.poetry]
name = "python-singerlake"
version = "0.1.0"
description = ""
authors = ["Ken Payne <ken@meltano.com>"]
readme = "README.md"
packages = [{ include = "singerlake", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pydantic = "^1.10.7"
filelock = "^3.11.0"
pyfarmhash = "^0.3.2"
numpy = "^1.24.2"
base58 = "^2.1.1"

[tool.poetry.group.dev.dependencies]
tox = "^4.4.12"
pytest = "^7.3.0"
black = "^23.3.0"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
